Fixed: (Indexers) Rate limit Download requests by Indexer
